Enable new shortlist behaviour for scoring
The default behaviour of shortlist performed an index select to retain
entries corresponding to shortlist candidates. This is desired for
decoding, where this reduction in the size of tensors offers an increase
in performance.

The new behaviour retains the full size of the tensors, and is designed
to be used with marian scorer. The list of indices for shortlist
candidates can be later used during loss computation.