Implement cross_entropy_shortlist operation
The cross_entropy_shortlist operation implements cross-entropy with a
modified softmax stage. This modified softmax uses the shortlist indices
to define the subset over which the softmax is normalized.

The motivation is to have entries inside the shortlist with the result
they would have in the absence of non-shortlist entries. This should
offer some comparison between results in scoring and decoding modes.