Merged PR 27976: Introduce new layer framework into master
Introduces the new layer framework into Marian master. This is currently not used anywhere for the existing models unless explicitly asked for. This also shouldn't modify any major existing functionality. The goal of this PR is to have the new code in master and use it for new things instead of the old code.

FYI: @<Roman Grundkiewicz> @<Hieu Hoang> the files in `src/layers_new` are the new framework. The rest is mostly unchanged apart from small modifications that allow for interaction with the new code. For now it exists in parallel to the old code.