quote CPUINFO in cmake
Without these quotes, cmake fails in a confusing manner
on systems whose cpuinfo output includes spaces.

This arose in the context of attempting to compile natively on an m1 mac.

$ /usr/sbin/sysctl -n machdep.cpu.features machdep.cpu.leaf7_features
sysctl: unknown oid 'machdep.cpu.leaf7_features'

Obviously, this didn't work out well; there is still much more to do.
Still, the quotes are cheap and eliminate a confusing failure mode.
For this reason, I added them to the linux as well as the darwin path.